Move sink responsibility to connection.
